Refactor decks.ts into separate files to separate guards, queries, mutations, schema, and seed data for decks. Also, add full-text and vector search indices.
